separation of modapi files into server/client/shared
